fix: A lot of fixes realted to the orders component

+++ database/migrations/2024_03_07_143031_create_order_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->bigIncrements('id');
             $table->foreignId('order_id')->constrained(table: 'orders')->cascadeOnDelete();
             $table->morphs('product_belongs_to'); /* This column can either belong to "products" or "products_by_buyers" */
-            $table->integer('product_price');
+            $table->float('product_price');
             $table->integer('product_qty');
             $table->tinyInteger('user_choice')->nullable()->comment('1-Alternative product that does the job, 2-Remove only this product from order, 3-Search for product in other stores, 4-Request a call from the store, 5-Cancel the order');
             $table->timestamps();
